Implement serialization for EmployeeList and CustomerList classes
Created a LoginUI and LoginCntl for Users to log in to the application. Test users have been
created and are written to at startup. Minimal functionality has been implemented for now.
The LoginUI will only check if the user is registered in the system, what type of user they
are, and will check to see if their login credentials are valid. Will let the user know if
they didn't select what type of user they are, or if their password is incorrect.